+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to the Aptos TypeScript SDK will be captured in this file. T
 
 # Unreleased
 
+- Add `AccountUtils` class to help with account serialization and deserialization
+- Add `SingleKeySigner` interface which adds the ability to get the `AnyPublicKey` from a `SingleKeyAccount`
+
 # 1.33.1 (2024-11-28)
 
 - Add `gasProfile` function to `Move` class to allow for gas profiling of Aptos Move functions